# openapi.json
{
  "components": {
    "schemas": {
...
      "AdminMe": {
        "additionalProperties": false,
        "properties": {
          "admin": {
            "additionalProperties": false,
            "properties": {
              "email": {
                "type": "string"
              },
              "id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "email"
            ],
            "type": "object"
          },
          "session": {
            "$ref": "#/components/schemas/SessionInfo"
          }
        },
        "required": [
          "admin",
          "session"
        ],
        "type": "object"
      },
...
      "ChangePasswordRequestInput": {
        "properties": {
          "currentPassword": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "newPassword": {
            "description": "Nova senha: de 8 a 128 caracteres.",
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Código estável, em inglês (ex.: `NOT_FOUND`, `TAB_ALREADY_CLOSED`).",
                "example": "VALIDATION_FAILED",
                "type": "string"
              },
              "details": {
                "additionalProperties": {},
                "description": "Dados extras do erro; em `VALIDATION_FAILED`, `{ fields: ValidationIssue[] }`.",
                "propertyNames": {
                  "type": "string"
                },
                "type": "object"
              },
              "message": {
                "description": "Mensagem em português, pode ser mostrada ao usuário.",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ForgotPasswordRequestInput": {
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminSession": {
        "description": "Sessão do admin da plataforma, definida pelo login do admin.",
        "in": "cookie",
        "name": "__Host-varal_admin_at",
        "type": "apiKey"
      },
      "panelSession": {
        "description": "Sessão do app dos clientes (dono e colaborador), definida pelo login.",
        "in": "cookie",
        "name": "__Host-varal_at",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "REST API do Varal (`/api/v1`). Contrato consumido pelos apps (spec 01, seção 3.1).",
    "title": "Varal API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/auth/logout": {
      "post": {
        "operationId": "AdminAuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sessão encerrada (ou já não havia sessão); cookies apagados."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Encerra a sessão do admin neste aparelho",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/admin/auth/me": {
      "get": {
        "operationId": "AdminAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMe"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`UNAUTHENTICATED`: sem sessão do admin ou sessão encerrada."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Admin logado",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/admin/auth/password/change": {
      "post": {
        "operationId": "AdminAuthController_change",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            },
            "description": "Senha trocada; as outras sessões foram encerradas e este aparelho recebeu uma sessão nova."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`WRONG_CURRENT_PASSWORD` ou `VALIDATION_FAILED`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`UNAUTHENTICATED`: sem sessão do admin ou sessão encerrada."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`RATE_LIMITED`: mais de 10 requisições em 900 s deste IP."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Troca a senha do admin (logado)",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/admin/auth/password/forgot": {
      "post": {
        "operationId": "AdminAuthController_forgot",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            },
            "description": "Sempre a mesma resposta, exista ou não o e-mail (RN-01.03)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`RATE_LIMITED`: mais de 5 requisições em 900 s deste IP."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Admin pede o link de redefinição de senha",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/admin/auth/password/reset": {
      "post": {
        "operationId": "AdminAuthController_reset",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Senha definida; todas as sessões do admin foram encerradas."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`INVALID_PASSWORD_TOKEN` ou `VALIDATION_FAILED`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`RATE_LIMITED`: mais de 10 requisições em 900 s deste IP."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Define a senha do admin com o token de um convite ou de uma redefinição",
        "tags": [
          "admin-auth"
        ]
      }
    },
    "/api/v1/admin/auth/refresh": {
      "post": {
        "operationId": "AdminAuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            },
            "description": "Novos cookies definidos."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`UNAUTHENTICATED`: sem token de renovação válido; cookies apagados."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Renova a sessão do admin",
        "tags": [
          "admin-auth"
        ]
      }
    },
...
    "/api/v1/auth/access-code/{code}": {
      "get": {
        "operationId": "AuthController_accessCode",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessCodeOrganization"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Código inexistente."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`RATE_LIMITED`: mais de 30 requisições em 60 s deste IP."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Nome da organização de um código de estabelecimento",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sessão encerrada (ou já não havia sessão); cookies apagados."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Encerra a sessão deste aparelho",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelMe"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`UNAUTHENTICATED`: sem sessão do app ou sessão encerrada."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "security": [
          {
            "panelSession": []
          }
        ],
        "summary": "Perfil, organização, unidades e estações permitidas",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/owner/login": {
      "post": {
        "operationId": "AuthController_ownerLogin",
        "parameters": [
          {
            "description": "UUID guardado no aparelho; obrigatório no login e gravado na sessão (spec 01, seção 7.2).",
            "in": "header",
            "name": "X-Device-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerLoginRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelMe"
                }
              }
            },
            "description": "Sessão aberta; cookies definidos."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Credenciais inválidas: sempre a mesma resposta, exista ou não o usuário."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`LOGIN_TEMPORARILY_LOCKED` (10 senhas erradas seguidas bloqueiam o identificador por 15 min) ou `RATE_LIMITED`."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Login do dono (e-mail e senha)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/change": {
      "post": {
        "operationId": "AuthController_change",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            },
            "description": "Senha trocada; as outras sessões foram encerradas e este aparelho recebeu uma sessão nova."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`WRONG_CURRENT_PASSWORD` ou `VALIDATION_FAILED`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`UNAUTHENTICATED`: sem sessão do app ou sessão encerrada."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`RATE_LIMITED`: mais de 10 requisições em 900 s deste IP."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "security": [
          {
            "panelSession": []
          }
        ],
        "summary": "Troca a senha (logado)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "operationId": "AuthController_forgot",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            },
            "description": "Sempre a mesma resposta, exista ou não o e-mail (RN-01.03)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`RATE_LIMITED`: mais de 5 requisições em 900 s deste IP."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Dono pede o link de redefinição de senha",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Senha definida; todas as sessões do usuário foram encerradas."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`INVALID_PASSWORD_TOKEN` ou `VALIDATION_FAILED` (senha com menos de 8 caracteres)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`RATE_LIMITED`: mais de 10 requisições em 900 s deste IP."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Define a senha com o token de um convite ou de uma redefinição",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            },
            "description": "Novos cookies definidos."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`UNAUTHENTICATED`: sem token de renovação válido; cookies apagados."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Renova a sessão (gira o token de renovação)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/staff/login": {
      "post": {
        "operationId": "AuthController_staffLogin",
        "parameters": [
          {
            "description": "UUID guardado no aparelho; obrigatório no login e gravado na sessão (spec 01, seção 7.2).",
            "in": "header",
            "name": "X-Device-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelMe"
                }
              }
            },
            "description": "Sessão aberta; cookies definidos."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Credenciais inválidas: sempre a mesma resposta, exista ou não o usuário."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`LOGIN_TEMPORARILY_LOCKED` (10 senhas erradas seguidas bloqueiam o identificador por 15 min) ou `RATE_LIMITED`."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Erro no formato `ErrorResponse` (spec 01, seção 5)."
          }
        },
        "summary": "Login do colaborador (código do estabelecimento, usuário e senha)",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [],
...
}